refactor: 중복되는 ControllerTest 및 fixture 를 global 패키지로 이동
