+++ src/main/resources/application.yml
@@ -2,7 +2,7 @@ spring:
   datasource:
     url: jdbc:mysql://localhost:3306/coupang
     username: root
-    password: jiwon0207
+    password:
     driver-class-name: com.mysql.cj.jdbc.Driver
 
 mybatis:
