+++ src/test/java/com/coupang/marketplace/global/fixture/UserFixture.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.fixture;
+package com.coupang.marketplace.global.fixture;
 
 import com.coupang.marketplace.user.domain.User;
