--- src/test/java/com/coupang/marketplace/product/controller/ControllerTest.java
@@ -1,37 +1,0 @@
-package com.coupang.marketplace.product.controller;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.filter.CharacterEncodingFilter;
-
-@ExtendWith(SpringExtension.class)
-@AutoConfigureMockMvc
-@SpringBootTest
-@Transactional
-public class ControllerTest {
-
-    @Autowired
-    protected MockMvc mvc;
-
-    @Autowired
-    protected WebApplicationContext ctx;
-
-    @Autowired
-    protected ObjectMapper objectMapper;
-
-    @BeforeEach
-    public void setUp() {
-        this.mvc = MockMvcBuilders.webAppContextSetup(ctx)
-                .addFilter(new CharacterEncodingFilter("UTF-8", true))
-                .build();
-    }
-}
