+++ src/test/java/com/coupang/marketplace/product/controller/ProductControllerTest.java
@@ -1,6 +1,7 @@
 package com.coupang.marketplace.product.controller;
 
 
+import com.coupang.marketplace.global.template.ControllerTestTemplate;
 import com.coupang.marketplace.global.util.MultiValueMapConverter;
 import com.coupang.marketplace.product.controller.dto.GetProductsRequest;
 import org.junit.jupiter.api.DisplayName;
@@ -14,7 +15,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class ProductControllerTest extends ControllerTest {
+public class ProductControllerTest extends ControllerTestTemplate {
 
     @DisplayName("로켓 배송을 선택하면 로켓 배송 가능 상품 목록을 보여준다.")
     @Test
