+++ src/test/java/com/coupang/marketplace/user/service/UserServiceTest.java
@@ -13,7 +13,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
-import static com.coupang.marketplace.fixture.UserFixture.*;
+import static com.coupang.marketplace.global.fixture.UserFixture.*;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
